requirements.txt
blinker==1.4
Flask==0.10.1
Flask-Script==2.0.5
Flask-Testing==0.4.2
Flask-WTF==0.12
funcsigs==0.4
httplib2==0.9.2
itsdangerous==0.24
Jinja2==2.8
linecache2==1.0.0
MarkupSafe==0.23
mock==1.3.0
pbr==1.8.1
pytz==2015.7
six==1.10.0
traceback2==1.4.0
twilio==6.9.0
unittest2==1.1.0
Werkzeug==0.11.3
wheel==0.24.0
WTForms==2.1
djangorestframework-3.11.0
